requirements.txt
torch
numpy
torchvision
huggingface_hub
pillow
